feat(discovery): refactor to use shared discovery abstraction from kork-core (#814)
