+++ gradle.properties
@@ -1,5 +1,5 @@
 fiatVersion=1.21.0
-korkVersion=7.50.0
+korkVersion=7.51.2
 org.gradle.parallel=true
 spinnakerGradleVersion=8.3.0
