+++ igor-web/src/test/groovy/com/netflix/spinnaker/igor/jenkins/JenkinsBuildMonitorSpec.groovy
@@ -27,6 +27,7 @@ import com.netflix.spinnaker.igor.jenkins.client.model.ProjectsList
 import com.netflix.spinnaker.igor.jenkins.service.JenkinsService
 import com.netflix.spinnaker.igor.polling.PollContext
 import com.netflix.spinnaker.igor.service.BuildServices
+import com.netflix.spinnaker.kork.discovery.DiscoveryStatusListener
 import com.netflix.spinnaker.kork.dynamicconfig.DynamicConfigService
 import org.slf4j.Logger
 import org.springframework.scheduling.TaskScheduler
@@ -53,7 +54,7 @@ class   JenkinsBuildMonitorSpec extends Specification {
             igorConfigurationProperties,
             new NoopRegistry(),
             new DynamicConfigService.NoopDynamicConfig(),
-            Optional.empty(),
+            new DiscoveryStatusListener(true),
             Optional.empty(),
             cache,
             buildServices,
